% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{add_sig}
\alias{add_sig}
\title{Convert p-values to significance notation}
\usage{
add_sig(x)
}
\arguments{
\item{x}{Numeric vector of p-values}
}
\value{
Character vector with significance levels: "ns", "*", "**", "***", "****"
}
\description{
Translate numeric p-values into standard significance notation
}
\examples{
add_sig(c(0.06, 0.03, 0.005, 0.0005, 0.00005))
# Returns: c("ns", "*", "**", "***", "****")
}
