\name{anotherFunction}
\alias{anotherFunction}
\title{Multiplies two numbers, adds a constant, and returns the result}
\usage{
  anotherFunction(a, b, c)
}
\arguments{
  \item{c}{A constant to add}

  \item{a}{A number I want to multiply by something}

  \item{b}{A multiplying constant}
}
\value{
  A number - the product of a and b, plus c
}
\description{
  Multiplies two numbers, adds a constant, and returns the
  result
}
\examples{
# Constants I want to multiply:
a <- 1
b <- 2
# Constant I want to add:
c <- 3
# Run anotherFunction:
anotherFunction(a,b,c)
}
\author{
  Paul Newcombe
}

