% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arimatex.R
\name{arimatex}
\alias{arimatex}
\title{Function to convert ARIMA results in a latex table}
\usage{
arimatex(model, tex = T, fixed = NULL, pr = F)
}
\arguments{
\item{model}{Arima model}

\item{tex}{logical, should output be in latex?}

\item{fixed}{vector indicating fixed parameters with NAs, default is NULL}

\item{pr}{logical, save output in file?}
}
\value{
model output
}
\description{
Function to convert ARIMA results in a latex table
}
\note{
built: 2013-09-30
}

