% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arimax2.R
\name{arimax2}
\alias{arimax2}
\title{This is a quick fix of the original arimax function which works for the
following cases: (0,0,0) model with transfer!  with this function, no fixed
parameters are possible!  last change: 2013-10-01 by Patrick Original author of
the arima function in R stats: Brian Ripley.  The arimax function is based on
the stats:::arima function, with modifications by Kung-Sik Chan.}
\usage{
arimax2(x, order = c(0, 0, 0), seasonal = list(order = c(0, 0, 0), period =
  NA), xreg = NULL, include.mean = TRUE, transform.pars = TRUE,
  fixed = NULL, init = NULL, method = c("CSS-ML", "ML", "CSS"), n.cond,
  optim.control = list(), kappa = 1e+06, io = NULL, xtransf,
  transfer = NULL)
}
\description{
This is a quick fix of the original arimax function which works for the
following cases: (0,0,0) model with transfer!  with this function, no fixed
parameters are possible!  last change: 2013-10-01 by Patrick Original author of
the arima function in R stats: Brian Ripley.  The arimax function is based on
the stats:::arima function, with modifications by Kung-Sik Chan.
}

