% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by.R
\name{by2df}
\alias{by2df}
\title{Convert by to data.frame}
\usage{
by2df(x, format = c("wide", "long", "array"), name = "output")
}
\arguments{
\item{x}{produced by stats::by}

\item{format}{type of output, or a formula passed to dcast}

\item{name}{assigned to the output variable}
}
\description{
Convert the output from the by function to a data frame
}
\details{
The stats::by function produces list-like objects, which can be converted to arrays or data frames
}
\examples{
data('meuse', package='sp')
byRes = by(meuse[,c('copper','lead')], meuse[, c('landuse','soil')], range)

dimnames(by2df(byRes, 'array'))
head(by2df(byRes, 'wide'))
head(by2df(byRes, 'long'))

if(requireNamespace('reshape2', quietly=TRUE)) {
	head(reshape2::dcast(by2df(byRes, 'long'), landuse ~ soil + output))
}


}
