% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera2.R
\name{camera2}
\alias{camera2}
\title{A wrapper around limma::camera() that returns the weights}
\usage{
camera2(
  y,
  index,
  design = NULL,
  contrast = ncol(design),
  weights = NULL,
  use.ranks = FALSE,
  allow.neg.cor = FALSE,
  inter.gene.cor = 0.01,
  trend.var = FALSE,
  sort = TRUE,
  ...
)
}
\arguments{
\item{expressionList}{a matrix of gene expression values}

\item{cameraMat}{matrix contain output from CAMERA}

\item{geneSets}{a list of list of geneIds}

\item{verbose}{Boolean whether to print out progress}
}
\value{
table of genes' mean expression values for each group
}
\description{
A wrapper around limma::camera() that returns the weights
}
