% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_compare.R
\name{category_compare}
\alias{category_compare}
\title{compares the mean expression between two groups of genes in a given
expression matrix, calculates the number and proportion of genes in the
input matrix that are up-regulated}
\usage{
category_compare(
  dat,
  setsIndices,
  grp1name = "group1",
  grp2name = "group2",
  grp1idx,
  grp2idx
)
}
\arguments{
\item{dat}{A matrix of gene expression}

\item{setsIndices}{A nested-list of gene indices for each GSEA category, most commonly made from limma::ids2indices()}

\item{grp1name}{String indicating name of the first group for the comparison}

\item{grp2name}{String indicating name of the second group for the comparison}

\item{grp1idx}{A vector of integer indices indicating columns belonging to group1}

\item{plot}{Boolean indicating whether to draw boxplots using ggplot2}
}
\value{
table of genes' mean expression values for each group
}
\description{
compares the mean expression between two groups of genes in a given
expression matrix, calculates the number and proportion of genes in the
input matrix that are up-regulated
}
\examples{
setsIndices <- ids2indices(geneIds, rownames(expressionMatrix))
categories <- names(setsIndices)
out <- category_compare(expressionMatrix,
                                  setIndices,
                                  grp1name = "control",
                                  grp2name = "experimental",
                                  which(grepl("control", colnames(expressionMatrix))),
                                  which(grepl("experimental", colnames(expressionMatrix)))
                                  )
}
