% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_with_dbscan.R
\name{cluster_with_dbscan}
\alias{cluster_with_dbscan}
\title{cluster using DBscan}
\usage{
cluster_with_dbscan(dat, epsilon, minpts = 5)
}
\arguments{
\item{dat}{data matrix}

\item{epsilon}{param to be passed to `dbscan()`}

\item{minpts}{param to be passed to `dbscan()`}
}
\description{
cluster using DBscan
}
