% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefTable.R
\name{coefTable}
\alias{coefTable}
\title{Matrix of coefficients}
\usage{
coefTable(x, maxChar = 6, approx = FALSE, ...)
}
\arguments{
\item{x}{model output from lm, glm, lmer, glmer, glmmTMB}

\item{maxChar}{truncate characters}

\item{approx}{if TRUE, create confidence intervals based on standard errors}

\item{...}{additional arguments for confint}
}
\value{
data frame with estimates and confidence interval.  coefficients are exponentiated if link is log or logit.
}
\description{
Produces a matrix of all coeficients suitable for printing
}
