% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{color.bar}
\alias{color.bar}
\title{Create color bar legend}
\usage{
color.bar(
  bc,
  min,
  max = -min,
  nticks = 5,
  ticks = seq(min, max, len = nticks),
  title = ""
)
}
\arguments{
\item{bc}{Vector of colors}

\item{min}{Minimum value}

\item{max}{Maximum value (default: -min)}

\item{nticks}{Number of ticks (default: 5)}

\item{ticks}{Tick positions}

\item{title}{Legend title (default: '')}
}
\value{
Base plot color bar
}
\description{
Create color bar legend
}
