% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{color.legend}
\alias{color.legend}
\title{Create color legend from formatted tibble}
\usage{
color.legend(color.df)
}
\arguments{
\item{color.df}{Tibble with 'values' and 'breaks' columns}
}
\value{
Base R barplot showing color legend
}
\description{
Create color legend from formatted tibble
}
