% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPASS_functions.R
\name{COMPASS_cell_counts}
\alias{COMPASS_cell_counts}
\title{extract cell counts from the COMPASSResult container (CR[[antigen]]$data$count_s)
!!! IMPORTANT: assumes data comes from time-course experiment and therefore has
joint_id (<ptid>_<time>)}
\usage{
COMPASS_cell_counts(CR)
}
\arguments{
\item{CR}{A COMPASSResult object}
}
\value{
table of cell counts, including null category
}
\description{
extract cell counts from the COMPASSResult container (CR[[antigen]]$data$count_s)
!!! IMPORTANT: assumes data comes from time-course experiment and therefore has
joint_id (<ptid>_<time>)
}
