% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{cor_hm}
\alias{cor_hm}
\title{Correlation heatmap with clustering}
\usage{
cor_hm(
  mat,
  title = "",
  x.str = "",
  y.str = "",
  color1 = "black",
  color2 = "white",
  color3 = "purple",
  cex_r = 1,
  cex_c = 0.6,
  mar = c(12, 13)
)
}
\arguments{
\item{mat}{Data matrix}

\item{title}{Plot title (default: "")}

\item{x.str}{X-axis label (default: "")}

\item{y.str}{Y-axis label (default: "")}

\item{color1}{Low correlation color (default: "black")}

\item{color2}{Mid correlation color (default: "white")}

\item{color3}{High correlation color (default: "purple")}

\item{cex_r}{Row label size (default: 1)}

\item{cex_c}{Column label size (default: 0.6)}

\item{mar}{Margins (default: c(12,13))}
}
\value{
Heatmap plot
}
\description{
Correlation heatmap with clustering
}
