% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{corr_across}
\alias{corr_across}
\title{Correlate features across a predictor variable}
\usage{
corr_across(feature.matrix, predictor.array)
}
\arguments{
\item{feature.matrix}{Matrix of feature values}

\item{predictor.array}{Predictor variable array}
}
\value{
Tibble with correlation statistics (R, p, p.adj)
}
\description{
Calculate Pearson correlations between features and a predictor
}
