% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{DEF_boxplot}
\alias{DEF_boxplot}
\title{Boxplot for differentially expressed features}
\usage{
DEF_boxplot(data, DE, feature, grouping)
}
\arguments{
\item{data}{Data frame}

\item{DE}{Output from comp_means}

\item{feature}{Feature name to plot}

\item{grouping}{Grouping variable name}
}
\value{
ggplot object
}
\description{
Create violin + boxplot with p-value annotation
}
\seealso{
\code{\link{DEF_boxplot_sig}}, \code{\link{comp_means}}

\code{vignette("comp_means")}
}
