% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{DEF_boxplot_sig}
\alias{DEF_boxplot_sig}
\title{Boxplot with significance stars}
\usage{
DEF_boxplot_sig(data, DE, feature, grouping)
}
\arguments{
\item{data}{Data frame}

\item{DE}{Output from comp_means}

\item{feature}{Feature name}

\item{grouping}{Grouping variable}
}
\value{
ggplot object
}
\description{
Variant of DEF_boxplot showing significance stars instead of p-values
}
\seealso{
\code{\link{DEF_boxplot}}

\code{vignette("comp_means")}
}
