% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diagtex.R
\name{diagtex}
\alias{diagtex}
\title{Function to display some diagnostics for (f)ECM model list in a latex table
built: 2013-10-29; Patrick Kraft}
\usage{
diagtex(models, tex = T, pr = F,
  caption = "Diagnostics and Fit Statistics", label = NULL)
}
\arguments{
\item{models}{(f)ECM model list}

\item{tex}{logical, should output be in latex}

\item{pr}{logical, should output be saved in new file}

\item{caption}{Caption of table}

\item{label}{Latex label}
}
\value{
model output
}
\description{
Function to display some diagnostics for (f)ECM model list in a latex table
built: 2013-10-29; Patrick Kraft
}

