% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{earliest}
\alias{earliest}
\title{Get earliest values from time series}
\usage{
earliest(dates, values)
}
\arguments{
\item{dates}{Vector of dates}

\item{values}{Corresponding values}
}
\value{
Mean of values at earliest date
}
\description{
Get earliest values from time series
}
\examples{
earliest(c(1, 2, 2, 3), c(10, 20, 30, 40)) # Returns 10
}
