% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECI.R
\name{ECI}
\alias{ECI}
\title{Calculate the Economic Complexity Index (ECI)}
\usage{
ECI(data, description = c("reg_code", "tech_class"), scale = "scale")
}
\arguments{
\item{data}{A (named) adjacency matrix of specialisations, with location in rows and activities in columns. These can be specialisations according to RCA or any other metric of specialisation that yields values that signal specialisations of locations in activites. This might be discreet (e.g. RCA, RSI) or real-valued (e.g. NRCA). See EconGeo::RCA e.g. for RCA or other functions in this package for other metrics of specialisations. See also Pintar and Essletzbichler (2022).}

\item{description}{Two strings indicating the name of dimensions (locations and activities, in this order), e.g. reg_code and tech_class (default).}

\item{scale}{A string indicating if the results should be scaled ("scale") between 0 and 100 or left as is.}
}
\value{
A list with tibble dataframes of complexity values for locations (ECI) and activities (PCI)
}
\description{
Calculate the ECI as in Hidalgo and Hausman (2009) but using the eigenvector method introduced by Caldarelli et al. (2012).
I also adapt the ECI in the following way:
\itemize{
\item I first calculate ECI and PCI as in Cardarelli et al. (2012)
\item Because the direction of an eigenvector is not defined, it might be the case that we get "flipped" results where the ECI we get is actually -ECI (from what we expect)
\item To remedy this, I calculate the pearson correlation of PCI with the ubiquity of the activity (product etc.). In case we get a positive correlation (which is against theoretical expectations), we multiply the eigenvector with -1 to "flip" the result
\item After that, I calculate the spearman correlation between PCI rank (of activities) and the ECI rank. This needs to be positive. If not, I "flip" the ECI
}
}
