\name{eqnarray}
\alias{eqnarray}
\title{Markdown equation array}
\usage{
eqnarray(x, mdToTex = FALSE)
}
\arguments{
\item{x}{a string with latex code for the equation}

\item{mdToTex}{if \code{FALSE} convert to markdown table, return \code{x} otherwise}
}
\description{
Multiline equation in a markdown table
}
\details{
A multi-line equation in an eqnarray or align document is converted to markdown code suitable for conversion to docx.
}
\examples{
}
