% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPASS_functions.R
\name{extract_COMPASS_cell_proportions}
\alias{extract_COMPASS_cell_proportions}
\title{calculate cell proportions from the COMPASSResult container (CR[[antigen]]$data$count_s)
!!! IMPORTANT: assumes data comes from time-course experiment and therefore has
joint_id (<ptid>_<time>)}
\usage{
extract_COMPASS_cell_proportions(CR)
}
\arguments{
\item{CR}{A COMPASSResult object}
}
\value{
table cell proportions, null category is included
}
\description{
calculate cell proportions from the COMPASSResult container (CR[[antigen]]$data$count_s)
!!! IMPORTANT: assumes data comes from time-course experiment and therefore has
joint_id (<ptid>_<time>)
}
