% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPASS_functions.R
\name{extract_COMPASS_cell_proportions_nonull}
\alias{extract_COMPASS_cell_proportions_nonull}
\title{calculate cell proportions from the COMPASSResult container (CR[[antigen]]$data$count_s)
!!! IMPORTANT: assumes data comes from time-course experiment and therefore has
joint_id (<ptid>_<time>)}
\usage{
extract_COMPASS_cell_proportions_nonull(CR)
}
\arguments{
\item{CR}{A COMPASSResult object}
}
\value{
table cell proportions, null category is NOT included
}
\description{
This function calculates percentages EXCLUDING THE NULL CATEGORY
eg. you have 10000 cells
5 cells express IL2 only, 3 cells express IFNg only, 2 cells express both (10 cells express any)
the percentages returned by this fx are IL2 = 5/10 = 0.5, IFNg = 3/10 = 0.3, IL2+IFNg = 2/10 = 0.2
this is different than when the calculation is over ALL cells (10000 in this example)
}
