% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPASS_functions.R
\name{extract_degree_from_COMPASS}
\alias{extract_degree_from_COMPASS}
\title{extract polyfunctionality degree from a COMPASSResult object}
\usage{
extract_degree_from_COMPASS(CR)
}
\arguments{
\item{CR}{A COMPASSResult object}
}
\value{
table of polyfunctionality degree
}
\description{
extract polyfunctionality degree from a COMPASSResult object
}
