% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPASS_functions.R
\name{extract_degree_from_melted_counts}
\alias{extract_degree_from_melted_counts}
\title{extract polyfunctionality degree from a table of melted counts
column names must be Boolean expression of cytokines (eg. "IL2&!IFNg&TNFa"...)}
\usage{
extract_degree_from_melted_counts(melted_counts)
}
\arguments{
\item{melted_counts}{A melted data.table of cell counts}
}
\value{
table of cell counts
}
\description{
extract polyfunctionality degree from a table of melted counts
column names must be Boolean expression of cytokines (eg. "IL2&!IFNg&TNFa"...)
}
