% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lm_pval.R
\name{extract_lm_pval}
\alias{extract_lm_pval}
\title{extracts F-statistic from a linear model object
then calculates P-value}
\usage{
extract_lm_pval(model)
}
\value{
pvalues
}
\description{
extracts F-statistic from a linear model object
then calculates P-value
}
\examples{
pvals <- extract_lm_pval(myModelObj)
}
