% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tsne_rectangle.R
\name{extract_tsne_rectangle}
\alias{extract_tsne_rectangle}
\title{plots the output of a tsne run on ICS flow data, highlights user-specified window, and return 
cell events within that window. Also perform t.test() on extracted events}
\usage{
extract_tsne_rectangle(
  dat,
  x_min,
  x_max,
  y_min,
  y_max,
  antigen,
  marker,
  polyDegree = 0,
  facet,
  title
)
}
\arguments{
\item{dat}{RTsne output matrix}

\item{x_min}{coordinate of bounding box to extract events}

\item{x_max}{coordinate of bounding box to extract events}

\item{y_min}{coordinate of bounding box to extract events}

\item{y_max}{coordinate of bounding box to extract events}

\item{antigen}{the antigen to subset, must exist in `dat`}

\item{polyDegree}{ICS polyfunctionality degree of data to extract}

\item{facet}{metadata to facet on, must exist in `dat`}

\item{title}{plot title}
}
\description{
plots the output of a tsne run on ICS flow data, highlights user-specified window, and return 
cell events within that window. Also perform t.test() on extracted events
}
