% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_compare.R
\name{gene_compare}
\alias{gene_compare}
\title{compares the mean expression between two groups of genes in a given
expression matrix, optionally makes boxplots of the comparison, faceted by gene
used in a pipeline, so checks are pretty rudimentary}
\usage{
gene_compare(
  dat,
  geneSet,
  grp1name = "group1",
  grp2name = "group2",
  grp1idx,
  grp2idx,
  plot = FALSE
)
}
\arguments{
\item{dat}{A matrix of gene expression}

\item{geneSet}{A vector of gene names in the gene expression matrix above}

\item{grp1name}{String indicating name of the first group for the comparison}

\item{grp2name}{String indicating name of the second group for the comparison}

\item{grp1idx}{A vector of integer indices indicating columns belonging to group1}

\item{plot}{Boolean indicating whether to draw boxplots using ggplot2}
}
\value{
table of genes' mean expression values for each group
}
\description{
compares the mean expression between two groups of genes in a given
expression matrix, optionally makes boxplots of the comparison, faceted by gene
used in a pipeline, so checks are pretty rudimentary
}
\examples{
out <- gene_compare(expressionMatrix
                   ,c("LAP3","NCOA7","IFIT2","STAT1","TRIM21")
                   ,grp1name = "control"
                   ,grp2name = "experimental"
                   ,which(grepl("control", colnames(expressionMatrix)))
                   ,which(grepl("experimental", colnames(expressionMatrix)))
                                  )
}
