% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_compare_with_weights.R
\name{gene_compare_with_weights}
\alias{gene_compare_with_weights}
\title{compares the mean expression between two groups of genes in a given
expression matrix, optionally makes boxplots of the comparison, faceted by gene
used in a pipeline, so checks are pretty rudimentary}
\usage{
gene_compare_with_weights(
  dat,
  camera2output,
  genesetidx,
  grp1name = "group1",
  grp2name = "group2",
  grp1idx,
  grp2idx
)
}
\arguments{
\item{dat}{A matrix of gene expression}

\item{camera2output}{output matrix from camera2()}

\item{genesetidx}{a list of lists of gene indices}

\item{grp1name}{String indicating name of the first group for the comparison}

\item{grp2name}{String indicating name of the second group for the comparison}

\item{grp1idx}{A vector of integer indices indicating columns belonging to group1}

\item{grp2idx}{A vector of integer indices indicating columns belonging to group1}
}
\value{
table of genes' mean expression values for each group
}
\description{
compares the mean expression between two groups of genes in a given
expression matrix, optionally makes boxplots of the comparison, faceted by gene
used in a pipeline, so checks are pretty rudimentary
}
\examples{
\dontrun{
out <- gene_compare(expressionMatrix
                   ,c("LAP3","NCOA7","IFIT2","STAT1","TRIM21")
                   ,grp1name = "control"
                   ,grp2name = "experimental"
                   ,which(grepl("control", colnames(expressionMatrix)))
                   ,which(grepl("experimental", colnames(expressionMatrix)))
                                  )
}
}
