% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{gene_OS_scatter}
\alias{gene_OS_scatter}
\title{Gene expression vs overall survival scatter plot}
\usage{
gene_OS_scatter(df, gene)
}
\arguments{
\item{df}{Data frame with time, status, and gene columns}

\item{gene}{Gene/feature name}
}
\value{
ggplot object
}
\description{
Gene expression vs overall survival scatter plot
}
\seealso{
\code{\link{coxph_all}}, \code{\link{KM_categorical}}

\code{vignette("survival_analysis")}
}
