% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{geom_vb}
\alias{geom_vb}
\title{Combine violin and box plots}
\usage{
geom_vb(box.width = 0.2, box.fill = "gray")
}
\arguments{
\item{box.width}{Width of boxplot (default: 0.2)}

\item{box.fill}{Fill color for boxplot (default: "gray")}
}
\value{
List of ggplot2 geom layers
}
\description{
Combine violin and box plots
}
