% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{get_lower_tri}
\alias{get_lower_tri}
\title{Get lower triangle of matrix}
\usage{
get_lower_tri(mat)
}
\arguments{
\item{mat}{Matrix}
}
\value{
Matrix with upper triangle set to NA
}
\description{
Extract lower triangle, useful for correlation heatmaps
}
\examples{
m <- matrix(1:9, 3, 3)
get_lower_tri(m) # Returns matrix with upper triangle as NA
}
