% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{get_upper_tri}
\alias{get_upper_tri}
\title{Get upper triangle of matrix}
\usage{
get_upper_tri(mat)
}
\arguments{
\item{mat}{Matrix}
}
\value{
Matrix with lower triangle set to NA
}
\description{
Extract upper triangle, useful for correlation heatmaps
}
\examples{
m <- matrix(1:9, 3, 3)
get_upper_tri(m) # Returns matrix with lower triangle as NA
}
