% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitrHooks.R
\name{hook_plot_htmlsubfig}
\alias{hook_plot_beamer}
\alias{hook_plot_htmlsubfig}
\alias{hook_plot_margins}
\title{Hooks for knitr}
\usage{
hook_plot_htmlsubfig(x, options)

hook_plot_beamer(x, options)

hook_plot_margins(before, options, envir)
}
\arguments{
\item{x}{see \code{\link[knitr]{hook_pdfcrop}}}

\item{options}{see \code{\link[knitr]{hook_pdfcrop}}}

\item{before}{see \code{\link[knitr]{hook_pdfcrop}}}

\item{envir}{see \code{\link[knitr]{hook_pdfcrop}}}
}
\description{
Hooks for including in knitr documents
}
\details{
\code{hook_plot_beamer} produces multi-plot figures in columns for beamer documents
\code{hook_plot_htmlsubfig} produces multi-plot figures in pipe tables for markdown documents
\code{hook_plot_margins} sets small plot margins, which must be set as a hook named \code{margins}.  The \code{margins} option 
in a chunk is an integer which scales the size of the margins.
}
\examples{
\dontrun{knit_hooks$set(plot=Pmisc::hook_plot_beamer)} 
\dontrun{knit_hooks$set(plot=Pmisc::hook_plot_htmlsubfig)} 
\dontrun{knit_hooks$set(margins = Pmisc::hook_plot_margins)}
\dontrun{opts_chunk$set(margins=1, fig.width=5, fig.height=3)}

}

