% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tides.R
\name{interpolate_tide}
\alias{interpolate_tide}
\title{Interpolate a tide level from a date}
\usage{
interpolate_tide(datetime, td)
}
\arguments{
\item{datetime}{A \code{\link{POSIXt}} datetime object, or a sensible character representation of this.}

\item{td}{Tide data. As returned by \code{\link{read_linz_tides}}}
}
\value{
A numeric vector of length one. Tide level in meters, usually.
}
\description{
Interpolate a tide level from a date
}
\examples{
td <- read_linz_tides("https://static.charts.linz.govt.nz/tide-tables/maj-ports/csv/Spit\%20Wharf\%202021.csv")
interpolate_tides("2021-01-01 05:42", td = td) # 1.90 (an exact match)
interpolate_tides("2021-01-01 05:30", td = td) # 1.896 (calculated)

with(td[1:8,], plot(rdt, level, type = "l"))
newx <- seq(from = td$rdt[1], to = td$rdt[8], length.out = 200)
lines(newx, interpolate_tides(newx, td = td), col = 2)

}
