% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tides.R
\name{interpolate_tides}
\alias{interpolate_tides}
\title{Interpolate tide levels from dates}
\usage{
interpolate_tides(datetimes, td)
}
\arguments{
\item{datetimes}{A vector of \code{\link{POSIXt}} datetime objects, or sensible character representations of these.}

\item{td}{Tide data. As returned by \code{\link{read_linz_tides}}}
}
\value{
A numeric vector equal in length to \code{datetimes}. Tide level in meters, usually.
}
\description{
Interpolate tide levels from dates
}
\examples{
td <- read_linz_tides("https://static.charts.linz.govt.nz/tide-tables/maj-ports/csv/Spit\%20Wharf\%202021.csv")
dt <- c("2021-01-01 05:42", "2021-03-01 17:59", "2021-01-01 05:30", "2021-03-01 19:00")
interpolate_tides(datetimes = dt, td = td)
# 1.900 2.000 1.896 1.877

}
