% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{KM_categorical}
\alias{KM_categorical}
\title{Kaplan-Meier plot by gene expression}
\usage{
KM_categorical(df, gene, time_scale, reclass = TRUE)
}
\arguments{
\item{df}{Data frame with time, status, and gene columns}

\item{gene}{Gene/feature name}

\item{time_scale}{Label for time axis}

\item{reclass}{Whether to split by median (default: TRUE)}
}
\value{
ggsurvfit object
}
\description{
Kaplan-Meier plot by gene expression
}
\seealso{
\code{\link{coxph_all}}, \code{\link{gene_OS_scatter}}

\code{vignette("survival_analysis")}
}
