% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{latest}
\alias{latest}
\title{Get latest values from time series}
\usage{
latest(dates, values)
}
\arguments{
\item{dates}{Vector of dates}

\item{values}{Corresponding values}
}
\value{
Mean of values at latest date
}
\description{
Get latest values from time series
}
\examples{
latest(c(1, 2, 2, 3), c(10, 20, 30, 40)) # Returns 40
}
