% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profL.R
\name{likSurface}
\alias{likSurface}
\title{Likelihood surfaces}
\usage{
likSurface(fit, se = seq(0, 3, len = 10))
}
\arguments{
\item{fit}{model fit from, for example, glm}

\item{se}{number of standard errors from MLE of each parameter}
}
\value{
List which includes an array of likelihoods
}
\description{
Evalues multivariable profile likelihoods
}
\examples{
    # from glm  help file
    counts <- c(18,17,15,20,10,20,25,13,12)
    outcome <- gl(3,1,9)
    treatment <- gl(3,3)
    glm.D93 <- glm(counts ~ outcome, family = poisson())
x = likSurface(glm.D93)
image(x$parameters[,1], x$parameters[,2], apply(x$logLik, c(2,3), min))

}
