% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lookfor.R
\name{lookfor}
\alias{lookfor}
\title{Function that allows searching for variable names in a dataset.  it shows all
variable names that are fit to the (truncated) search entry and reports whether
the item list was complete and exact (i.e. not truncated) built: 2013-09-28,
Patrick Kraft}
\usage{
lookfor(data, term, summarize = F)
}
\arguments{
\item{data}{data frame}

\item{term}{term to look for in variable names}

\item{summarize}{logical, include summary statistics for variables?}
}
\value{
variable list
}
\description{
Function that allows searching for variable names in a dataset.  it shows all
variable names that are fit to the (truncated) search entry and reports whether
the item list was complete and exact (i.e. not truncated) built: 2013-09-28,
Patrick Kraft
}

