% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{lu}
\alias{lu}
\title{Length of unique values}
\usage{
lu(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector}

\item{na.rm}{Logical, whether to remove NA values before counting (default: FALSE)}
}
\value{
Integer count of unique values
}
\description{
Calculate the number of unique values in a vector
}
\examples{
lu(c(1, 1, 2, 3, 3, 3))
lu(c("a", NA, "b", "c"), na.rm = TRUE) # Returns 3
}
