% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{make_grid}
\alias{make_grid}
\title{Arrange multiple ggplot objects in a grid}
\usage{
make_grid(
  ggplots_list,
  nrow = floor(sqrt(length(ggplots_list))),
  ncol = ceiling(length(ggplots_list)/nrow)
)
}
\arguments{
\item{ggplots_list}{List of ggplot objects}

\item{nrow}{Number of rows (default: auto-calculated)}

\item{ncol}{Number of columns (default: auto-calculated)}
}
\value{
Arranged plot grid
}
\description{
Arrange multiple ggplot objects in a grid
}
