% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gsea_overlap_igraph.R
\name{make_gsea_overlap_igraph}
\alias{make_gsea_overlap_igraph}
\title{this function takes a bioConductor EList (eg. voom-transformed RNASeq counts)
GSEA output matrix from limma's camera() and a user-specified geneSet
and create an igraph graph where the nodes are gene lists and the edge weight
is how many expressed genes are common between the gene lists
Phu T. Van, FHCRC 2017, w/ substantial help and input from C.Murie & V.Voillet}
\usage{
make_gsea_overlap_igraph(expressionList, cameraMat, geneSets, verbose = FALSE)
}
\arguments{
\item{plot}{Boolean indicating whether to draw boxplots using ggplot2}
}
\value{
table of genes' mean expression values for each group
}
\description{
this function takes a bioConductor EList (eg. voom-transformed RNASeq counts)
GSEA output matrix from limma's camera() and a user-specified geneSet
and create an igraph graph where the nodes are gene lists and the edge weight
is how many expressed genes are common between the gene lists
Phu T. Van, FHCRC 2017, w/ substantial help and input from C.Murie & V.Voillet
}
\examples{
vDat <- voom(exprs(eDat), design=designMat, plot=FALSE, lib.size=libNorm)
}
