% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdTable.R
\name{mdTable}
\alias{mdTable}
\title{Markdown tables}
\usage{
mdTable(x, col.just = "r", guessGroup = FALSE, ..., mdToTex = "auto")
}
\arguments{
\item{x}{a matrix or data frame}

\item{col.just}{horizontal justification of text in cells}

\item{guessGroup}{attempt to guess cgroup and rgroup, for reshape2 output}

\item{...}{other arguments for \code{\link[Hmisc]{latex}}, \code{\link[htmlTable]{htmlTable}} or \code{\link[knitr]{kable}}.}

\item{mdToTex}{if \code{FALSE} convert to markdown table, return \code{x} otherwise}
}
\value{
character string which knitr prints 'asis'
}
\description{
Markdown tables with pandoc-crossref support
}
\details{
\code{\link[Hmisc]{latex}} is called if \code{mdToTex} is \code{TRUE}.  Otherwise, 
\code{\link[htmlTable]{htmlTable}} is called if arguments such as \code{cgroup} are present
and \code{\link[knitr]{kable}} is called to produce a markdown table.

Captions with labels suitable for pandoc-crossref are added.  If not provided, the
label is \code{tbl:} followed by the chunk label.

When \code{mdToTex} is \code{'auto'} (the default), it is set to
\code{any(commandArgs()=='mdToTex', na.rm=TRUE)}
}
\examples{
mytable = as.data.frame(matrix(runif(6),,3))
names(mytable) = month.name[1:ncol(mytable)]

cat(mdTable(x=mytable, digits=3,
  caption = 'the table', caption.loc='bottom',
mdToTex=FALSE))

}
