% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metro_converter.r
\name{metro_converter}
\alias{metro_converter}
\title{Convert NUTS codes to Metropolitan regions}
\usage{
metro_converter(data)
}
\arguments{
\item{data}{A list of NUTS3 codes that should be translated to metro regions, make sure you don't leave out codes that belong to metro regions}
}
\value{
A dataframe with converted nuts codes to metro codes
}
\description{
Convert NUTS3 (NUTS 2013) regions to metropolitan regions (2013)
}
