% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{named_num_to_df}
\alias{named_num_to_df}
\title{Convert named numeric vector to data frame}
\usage{
named_num_to_df(named_numeric)
}
\arguments{
\item{named_numeric}{A named numeric vector}
}
\value{
Tibble with names and values columns
}
\description{
Convert named numeric vector to data frame
}
