% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuts.converter.r
\name{nuts.converter}
\alias{nuts.converter}
\title{Convert NUTS classifications}
\usage{
nuts.converter(
  data,
  from = "2016",
  to = "2013",
  weight = tibble::tibble("equal", "dummy"),
  merged = "sum"
)
}
\arguments{
\item{data}{A dataframe that includes regional data, codes in first column, data in second column, all other columns after second
will be retained but not translated (just copied)}

\item{from}{Text that identifies the nuts classification the data comes in ("2013" or "2016")}

\item{to}{Text that identifies the nuts classification the data should be translated to ("2013" or "2016")}

\item{weight}{A dataframe that gives a weight for target NUTS codes which are split, defaults to equal weighing,
optional set first element 'same' to give all split regions same value as old region (appropriate for relative variables)}

\item{merged}{Either "sum" or "average" depending on what is appropriate for merged regions}
}
\value{
A dataframe with converted nuts classification
}
\description{
Convert data between 2013 and 2016 nuts classification, only NUTS3 so far and
only works for one variable at a time
}
