% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outWidth.R
\name{out.width}
\alias{out.width}
\title{Output width for knitr figures}
\usage{
out.width(x, mdToTex = "auto")
}
\arguments{
\item{x}{width, fraction of page width}

\item{mdToTex}{if \code{FALSE} convert to markdown table, return \code{x} otherwise}
}
\description{
Creates output width strings for html and latex
}
\examples{
Pmisc::out.width(0.5, 'latex')
Pmisc::out.width(0.5, 'markdown')
mdToTex=FALSE
Pmisc::out.width(0.5, 'auto')
mdToTex=TRUE
Pmisc::out.width(0.5, 'auto')

}
