% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{p.readjust}
\alias{p.readjust}
\title{Readjust p-values from comp_means output}
\usage{
p.readjust(cm_out, p.adjust_method = "fdr")
}
\arguments{
\item{cm_out}{Output from comp_means}

\item{p.adjust_method}{Adjustment method (default: "fdr")}
}
\value{
Data frame with updated p.adj columns
}
\description{
Readjust p-values from comp_means output
}
