% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpca.R
\name{plotpca}
\alias{plotpca}
\title{this is a convenience wrapper to make plotting PCA biplots a bit nicer
contribute tothe PCs but interferes with plotting
it also colors the groups with user-supplied labels}
\usage{
plotpca(df, title = NULL, labels = NULL, varname.adjust = 2)
}
\arguments{
\item{df}{data.frame of input data}

\item{title}{text to title the output plot}

\item{labels}{labels to label the groups}

\item{varname.adjust}{parameter to be passed to `ggbiplot`}
}
\value{
a ggplot2 plot
}
\description{
this is a convenience wrapper to make plotting PCA biplots a bit nicer
contribute tothe PCs but interferes with plotting
it also colors the groups with user-supplied labels
}
