% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{png_pdf}
\alias{png_pdf}
\title{Save plot as PNG and PDF}
\usage{
png_pdf(p, filename, width, height, res = 1000)
}
\arguments{
\item{p}{ggplot object}

\item{filename}{Filename without extension}

\item{width}{Width in inches}

\item{height}{Height in inches}

\item{res}{Resolution for PNG (default: 1000)}
}
\value{
List (side effect: saves files)
}
\description{
Save plot as PNG and PDF
}
