% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{rats}
\alias{rats}
\title{Function to transform an unordered rats dataset in the right format for R
built: 2013-09-30, Patrick Kraft}
\usage{
rats(data, items, byrow = T)
}
\arguments{
\item{data}{dataset loaded in R}

\item{items}{number of variables in dataser}

\item{byrow}{logical, is data organized by row?}
}
\value{
usable dataframe
}
\description{
Function to transform an unordered rats dataset in the right format for R
built: 2013-09-30, Patrick Kraft
}

