% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tides.R
\name{read_linz_tides}
\alias{read_linz_tides}
\title{Read LINZ-formatted tide information}
\usage{
read_linz_tides(filename)
}
\arguments{
\item{filename}{}
}
\value{
A dataframe representation of the LINZ tide data.
}
\description{
Read LINZ-formatted tide information
}
\examples{
# NB: Usually, it is best to download these files locally, rather than
# rely on the URL structure on the LINZ site remaining stable.
td <- read_linz_tides("https://static.charts.linz.govt.nz/tide-tables/maj-ports/csv/Spit\%20Wharf\%202021.csv")

}
