% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers based on IQR}
\usage{
remove_outliers(df, var, coef = 3)
}
\arguments{
\item{df}{Data frame}

\item{var}{Variable name to check for outliers}

\item{coef}{Coefficient multiplier for IQR (default: 3)}
}
\value{
Data frame with outliers removed
}
\description{
Remove points outside of coef*IQR from Q1/Q3
}
