% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{reorder_cormat}
\alias{reorder_cormat}
\title{Reorder correlation matrix by hierarchical clustering}
\usage{
reorder_cormat(cormat)
}
\arguments{
\item{cormat}{Correlation matrix}
}
\value{
Reordered correlation matrix
}
\description{
From STHDA, reorders correlation matrix for better visualization
}
