% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{rm_from_array}
\alias{rm_from_array}
\title{Remove specified values from array}
\usage{
rm_from_array(array, values_to_remove)
}
\arguments{
\item{array}{Input vector}

\item{values_to_remove}{Values to exclude}
}
\value{
Filtered vector
}
\description{
Tidy-compatible removal for use in piping operations
}
\examples{
rm_from_array(1:5, c(2, 4)) # Returns c(1, 3, 5)
}
