% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSI.R
\name{RSI}
\alias{RSI}
\title{Calculate RSI index (Menzel and Maicher, 2017)}
\usage{
RSI(data, binary = T, threshold = c("std", "1"), shift = F)
}
\arguments{
\item{data}{An incidence matrix with agent in rows and activities in columns.}

\item{binary}{Logical if index should be binarised with 1 signalling specialisation. Default is binary.}

\item{threshold}{Threshold for binarisation, Two strings; either standard deviations ("std") plus the factor with which std is multiplied (second element)
or "0" for regular binarising at zero cutoff. Default is "std","1".}

\item{shift}{Logical indicating whether the index should be shifted to range between 0 and 2 instead of -1 and 1}
}
\value{
A sparse matrix (agent x activities) with RSI values
}
\description{
Calculate RSI index which is an alternative but related specialisation index derived from incidence matrices
}
