% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDensHaz.R
\name{sampleDensHaz}
\alias{sampleDensHaz}
\title{Sample hazard functions}
\usage{
sampleDensHaz(fit, x, n = 1, covariates = c(`(Intercept)` = 1), scale = 1)
}
\arguments{
\item{fit}{result from inla}

\item{x}{sequence of times}

\item{n}{number of samples}

\item{covariates}{values of covariates to use for computing the hazard}

\item{scale}{scale the hazard (12 converts monthly hazard to yearly)}
}
\value{
matrix or array
}
\description{
Take posterior samples of hazard funnctions from an INLA fit
}
