% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{scm}
\alias{scm}
\title{Scale color manual for formatted color tibbles}
\usage{
scm(color.df)
}
\arguments{
\item{color.df}{Tibble with 'values' and 'breaks' columns}
}
\value{
ggplot2 scale_color_manual layer
}
\description{
Scale color manual for formatted color tibbles
}
