% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalFormula.R
\name{seasonalFormula}
\alias{seasonalFormula}
\title{Seasonal formula}
\usage{
seasonalFormula(period = 365.25, harmonics = 1:2, var = "x")
}
\arguments{
\item{period}{1/frequency, 365.25 for daily data with yearly cycles}

\item{harmonics}{multiplied by period, one sin/cosine pair per harmonic}

\item{var}{variable name}
}
\value{
a formula with sine/cosine pairs
}
\description{
A formula with sines and cosines
}
