% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{select_as_array}
\alias{select_as_array}
\title{Select column as array}
\usage{
select_as_array(df, col)
}
\arguments{
\item{df}{Data frame}

\item{col}{Column name to select}
}
\value{
Vector of values from specified column
}
\description{
Select column as array
}
