% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{sfm}
\alias{sfm}
\title{Scale fill manual for formatted color tibbles}
\usage{
sfm(color.df)
}
\arguments{
\item{color.df}{Tibble with 'values' and 'breaks' columns}
}
\value{
ggplot2 scale_fill_manual layer
}
\description{
Scale fill manual for formatted color tibbles
}
