% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{slope}
\alias{slope}
\title{Calculate slope of linear model}
\usage{
slope(x, y)
}
\arguments{
\item{x}{Numeric vector (predictor)}

\item{y}{Numeric vector (response)}
}
\value{
Numeric slope coefficient
}
\description{
Calculate average slope of y ~ x using linear regression
}
