% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_div.R
\name{struct_div}
\alias{struct_div}
\title{Calculate structural diversity as in Broekel (2019)}
\usage{
struct_div(
  data,
  step,
  cutoff = 10,
  subnetwork_number = 50,
  walktrap_steps = 150,
  graphlets = F,
  activity = "tech_class",
  activity_long = "tech_class_long",
  location = "appln_id",
  min_edges = 5,
  parallel = c("sequential", "1")
)
}
\arguments{
\item{data}{A dataframe activity and location data (e.g. techs on patents) for step 1. The output of the previous step (list)
for steps 2 and 3.}

\item{step}{An integer indicating which step of calculating structural diversity is to be done.
All three need to be executed in order and are prerequisites of the following step. 1 for creating combinatorial networks,
2 for creating subnetworks of combinatorial network on which step 3 will carry out final calculations for and report results.}

\item{cutoff}{An integer of the minimum number of locations (or patents e.g.) an activity is present to be included in the calculations.
Default is 10.}

\item{subnetwork_number}{A numeric setting the number of subnetworks that should be drawn from full combinatorial networks.
Default is 50, as in Broekel (2019).}

\item{walktrap_steps}{A numeric setting the number of steps the random walktrap should go when drawing the subnetworks.
Default is 150, as in Broekel (2019). Alternatively, a share (0,1) of the total number of nodes in the combinatorial network can be given. A maximum of 500 steps is set, when doing relative.}

\item{graphlets}{Logical indicating whether graphlets (T) (as in Broekel (2019)) or motifs (F) (as in Emmert-Streib, Dehmer 2012) should be used.
Default is motifs.}

\item{activity}{A string giving the variable name of activity in the data. Default is tech_class.}

\item{location}{A string giving the variable name of location in the data. Default is appln_id.}

\item{min_edges}{A numeric giving the minimum number of edges a combinatorial network needs to have to be considered. Default is 5.}

\item{parallel}{A string with two elements giving the preferred method for parallel computing (in step 2). "sequential", "multisession" (windows local pc)
and number of cores that should be engaged.}
}
\value{
A list with combinatorial network for each tech for step 1. A list with subnetworks and metrics for each tech for step 2.
A list with a tibble with unnested metrics of structural diversity for each tech (i) and subnetwork (j).
}
\description{
Calculate structural diversity index of activities from an activities dataframe where each activity needs to be
associated with a location.
}
