% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{survival_volcano}
\alias{survival_volcano}
\title{Volcano plot for survival analysis results}
\usage{
survival_volcano(coxph_all_out)
}
\arguments{
\item{coxph_all_out}{Output from coxph_all function}
}
\value{
ggplot object
}
\description{
Volcano plot for survival analysis results
}
\seealso{
\code{\link{coxph_all}}

\code{vignette("survival_analysis")}
}
