% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_eigenvector.R
\name{test_eigenvector}
\alias{test_eigenvector}
\title{Compute eigenvectors of matrix}
\usage{
test_eigenvector(mat)
}
\arguments{
\item{mat}{A quadratic test matrix}
}
\value{
a vector of eigenvectors
}
\description{
Function is just here to test how the creation of functions work
}
\examples{
mat <- matrix(sample(0:100,25,replace=T), ncol = 5)
rownames(mat) <- c ("R1", "R2", "R3", "R4", "R5")
colnames(mat) <- c ("I1", "I2", "I3", "I4","I5")
eigenvalues <- test_eigenvector(mat)
}
