% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{tibble_to_mat}
\alias{tibble_to_mat}
\title{Convert tibble to matrix with row names}
\usage{
tibble_to_mat(df, col.rownames)
}
\arguments{
\item{df}{A tibble or data frame}

\item{col.rownames}{Column name to use for row names}
}
\value{
Matrix with specified column as row names
}
\description{
Convert tibble to matrix with row names
}
