% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{title_format}
\alias{title_format}
\title{Format plot titles for publication}
\usage{
title_format(face = "bold", hjust = 0.5, size = 11)
}
\arguments{
\item{face}{Font face (default: "bold")}

\item{hjust}{Horizontal justification (default: 0.5)}

\item{size}{Font size (default: 11)}
}
\value{
ggplot2 theme layer
}
\description{
Format plot titles for publication
}
