% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniMultivariateComparison.R
\name{UniMultivariateComparison}
\alias{UniMultivariateComparison}
\title{Formats predictions into a ROCR `object' to plot using the RocPlots function.}
\usage{
UniMultivariateComparison(outcome, covariates, confounders = NULL, data,
  pretty = F, model = "linear")
}
\arguments{
\item{outcome}{Outcome variable}

\item{covariates}{Covariates to calculate univariate and multivariate effects for}

\item{confounders}{List of confounders to include in univariate and multivariate models}

\item{data}{Matrix or datafram containing the variables above.}

\item{pretty}{Whether to retrun a prettyified table (which contains strings so might be less useful).}

\item{model}{"logistic" or "linear"}
}
\description{
Runs univariate and multivariate analyses for a given set of covariates
}
\author{
Paul Newcombe
}
