% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{upset_plot}
\alias{upset_plot}
\title{Create upset plot from binary data}
\usage{
upset_plot(df, value = TRUE, cex.text = 1)
}
\arguments{
\item{df}{Data frame with binary/logical columns}

\item{value}{Value to consider as TRUE (default: TRUE)}

\item{cex.text}{Text size multiplier}
}
\value{
ggplot upset plot
}
\description{
Create upset plot from binary data
}
\seealso{
\code{\link{upset_side}}, \code{\link{make_grid}}

\code{vignette("upset_plots")}
}
