% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{volcano_cm}
\alias{volcano_cm}
\title{Create volcano plot from comp_means results}
\usage{
volcano_cm(comp_means_output, x = "mean_dif", max_overlaps = 10)
}
\arguments{
\item{comp_means_output}{Output from comp_means function}

\item{x}{X-axis variable: "mean_dif" or "log2fc"}

\item{max_overlaps}{Maximum label overlaps for ggrepel}
}
\value{
ggplot object
}
\description{
Create volcano plot from comp_means results
}
