% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmisc.R
\name{volcano_cm_labelall}
\alias{volcano_cm_labelall}
\title{Volcano plot with labels for all features}
\usage{
volcano_cm_labelall(comp_means_output, x = "mean_dif", max_overlaps = 10)
}
\arguments{
\item{comp_means_output}{Output from comp_means}

\item{x}{X-axis variable}

\item{max_overlaps}{Maximum label overlaps}
}
\value{
ggplot object
}
\description{
Variant of volcano_cm that labels all points regardless of significance
}
\seealso{
\code{\link{volcano_cm}}

\code{vignette("comp_means")}
}
