% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vratio.R
\name{vratio}
\alias{vratio}
\title{Function runs the variance ratio unit root test outlined in Cochrane (1988),
Diebold (1989) and Lo and McKinley (1988) Code is translatet from the original
(at)vratio procedure for RATS by Christopher J. Zorn built: 2013-10-28; Patrick
Kraft}
\usage{
vratio(series, lags = 2)
}
\arguments{
\item{series}{time series to test for unit root}

\item{lags}{number of lags, default is 2}
}
\value{
variance ratio test result
}
\description{
Function runs the variance ratio unit root test outlined in Cochrane (1988),
Diebold (1989) and Lo and McKinley (1988) Code is translatet from the original
(at)vratio procedure for RATS by Christopher J. Zorn built: 2013-10-28; Patrick
Kraft
}

