% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_one_run.R
\name{WGCNA_one_run}
\alias{WGCNA_one_run}
\title{perform a single run of the Weight Gene Correlation Network Analysis
(WGCNA, Langfelder 2008)
modified from original code by Carl Murie}
\usage{
WGCNA_one_run(
  dat,
  netType = "unsigned",
  pow = NULL,
  iter = 1,
  defaultPow = 3,
  showPlots = TRUE,
  ds = 1
)
}
\arguments{
\item{dat}{A matrix of gene expression}

\item{netType}{String indicating network type, default "unsigned"}

\item{pow}{Integer indicating power}

\item{iter}{Integer indicating number of iterations}

\item{defaultPow}{Integer indicating default power}

\item{showPlots}{Boolean indicating whether to plot}

\item{ds}{Integer between 0 and 4, tree cutting parameter}
}
\value{
table of cluster memberships
}
\description{
perform a single run of the Weight Gene Correlation Network Analysis
(WGCNA, Langfelder 2008)
modified from original code by Carl Murie
}
\examples{
\dontrun{
out <- WGCNA_one_run(dat, "unsigned", 1,1,3,TRUE,1)
}
}
