% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteDescription.R
\name{WriteDescription}
\alias{WriteDescription}
\title{Automates writing of the description file}
\usage{
WriteDescription(package.name, import.packages = NULL,
  package.location = "/Users/pauln/Dropbox/Work Projects/R Packages",
  version = 0.1)
}
\arguments{
\item{package.name}{The name of the package}

\item{package.location}{Directory all R packages are stored in.}

\item{version}{Version number which today's date will be appended to. Default is 0.1}
}
\description{
Automates writing of the description file for packages. The version number has today's date
appended. Takes descriptive information from the files TextTitle.txt, TextDescription.txt,
and TextOverview.txt which are stored in the root of the package folder.
}
