% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera2.R
\name{camera2}
\alias{camera2}
\title{A wrapper around limma::camera() that returns weights of the linear model}
\usage{
camera2(
  y,
  index,
  design = NULL,
  contrast = ncol(design),
  weights = NULL,
  use.ranks = FALSE,
  allow.neg.cor = FALSE,
  inter.gene.cor = 0.01,
  trend.var = FALSE,
  sort = TRUE,
  ...
)
}
\arguments{
\item{y}{a matrix of gene expression values}

\item{index}{`camera` parameter}

\item{design}{`camera` parameter}

\item{contrast}{`camera` parameter}

\item{weights}{`camera` parameter}

\item{use.ranks}{`camera` parameter}

\item{allow.neg.cor}{`camera` parameter}

\item{inter.gene.cor}{`camera` parameter}

\item{trend.var}{`camera` parameter}

\item{sort}{`camera` parameter}

\item{...}{other parameters to be passed to `camera()`}
}
\value{
list of 2 tables: summary of weights, and gene-level values
}
\description{
A wrapper around limma::camera() that returns weights of the linear model
}
