% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadIfOld}
\alias{downloadIfOld}
\title{Download a file}
\usage{
downloadIfOld(url, path = getwd(), file = file.path(path,
  make.names(basename(url))), age = "6 months", overwrite = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{url}{a string specifying the remote location}

\item{path}{local directory to store downloaded files}

\item{file}{local file name}

\item{age}{maximum age of the local file}

\item{overwrite}{replace existing unzipped files}

\item{verbose}{print additional information}

\item{...}{additional arguments for \code{\link[utils]{download.file}}}
}
\value{
A character string of files downloaded or unzipped.
}
\description{
Checks if a downloaded file is old and re-downloads is necessary
}
\details{
Checks if a downloaded file is old and re-downloads is necessary.  Any zip files are unzipped.
}
\examples{
\dontrun{  
  theFiles = Pmisc::downloadIfOld(
    'https://cran.r-project.org/src/base/Historic/Windows/mva.zip'
  )
  theFiles
}
}
\seealso{
\code{\link[utils]{download.file}}, \code{\link[utils]{unzip}}
}
