% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownEquation.R
\name{eqnarray}
\alias{eqnarray}
\title{Markdown equation array}
\usage{
eqnarray(x, mdToTex = "auto")
}
\arguments{
\item{x}{a string with latex code for the equation}

\item{mdToTex}{if \code{FALSE} convert to markdown table, return \code{x} otherwise}
}
\description{
Multiline equation in a markdown table
}
\details{
A multi-line equation in an eqnarray or align document is converted to markdown code suitable for conversion to docx.
}
\examples{
# backslashes may not appear correctly
cat(eqnarray(x=
      "\\x5cbegin{align*}
      Y_i = & N(u_i, t^2)\\x5c\\x5c
      u_i= & X_i B
      \\x5cend{align*}", 
    FALSE) )

}
