% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatRocrObjects.R
\name{FormatRocrObjects}
\alias{FormatRocrObjects}
\title{Formats predictions into a ROCR `object' to plot using the RocPlots function.}
\usage{
FormatRocrObjects(predictions.list, true.outcomes, fpr.stop = 1)
}
\arguments{
\item{predictions.list}{A named list of predictions for each different analysis/model,
for each person/variable, and for each fold/replicate. First level is the analysis/model
names. Second level can be matrices (rows people/vars, cols folds/replicates) or a list
where each element corresponds to a different fold/replicate, and contains a vector of
predictions across people/replicates.}

\item{true.outcomes}{The true outcomes of disease/variable causality. Can be a matrix
(rows people/vars, cols folds/replicates) or a list where each element corresponds to
a different fold/replicate, and contains a vector of truth indicators. NOTE: These are
assumed to be the same across analyses.}

\item{fpr.stop}{Maximum x to go up to on the x-axis for a partial ROC}
}
\description{
Formats predictions into a ROCR `object' to plot using the RocPlots function.
}
\author{
Paul Newcombe
}
