% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gsea_overlap_igraph.R
\name{make_gsea_overlap_igraph}
\alias{make_gsea_overlap_igraph}
\title{this function takes a bioConductor EList (eg. voom-transformed RNASeq counts)
GSEA output matrix from limma's camera() and a user-specified geneSet
and create an igraph graph where the nodes are gene lists and the edge weight
is how many expressed genes are common between the gene lists
Phu T. Van, FHCRC 2017, w/ substantial help and input from C.Murie & V.Voillet}
\usage{
make_gsea_overlap_igraph(expressionList, cameraMat, geneSets, verbose = FALSE)
}
\arguments{
\item{expressionList}{a matrix of gene expression values}

\item{cameraMat}{matrix contain output from CAMERA}

\item{geneSets}{a list of list of geneIds}

\item{verbose}{Boolean whether to print out progress}
}
\value{
table of genes' mean expression values for each group
}
\description{
this function takes a bioConductor EList (eg. voom-transformed RNASeq counts)
GSEA output matrix from limma's camera() and a user-specified geneSet
and create an igraph graph where the nodes are gene lists and the edge weight
is how many expressed genes are common between the gene lists
Phu T. Van, FHCRC 2017, w/ substantial help and input from C.Murie & V.Voillet
}
\examples{
\dontrun{
vDat <- voom(exprs(eDat), design=designMat, plot=FALSE, lib.size=libNorm)
res <- camera(vDat, setsIndices, design=designMat, contrast=cons[i], sort=TRUE)
geneSets <- geneIds(getGmt(gmtFile))
n <- make_gsea_igraph(vDat, res, geneSets)
plot(n, edge.width=E(n)$overlap*.1, vertex.color="white", vertex.label.cex=0.5)
}
}
