% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outWidth.R
\name{out.width}
\alias{out.width}
\title{Output width for knitr figures}
\usage{
out.width(x, ...)
}
\arguments{
\item{x}{width, fraction of page width, a number between zero and 1}

\item{...}{other arguments, currently ignored}
}
\description{
Creates output width strings for html and latex
}
\examples{
Pmisc::out.width(0.5)

}
