\docType{package}
\name{Pmisc-package}
\alias{Pmisc-package}
\title{All my miscellaneous functions.}
\description{
  All my miscellaneous functions.
}
\details{
  \tabular{ll}{ Package: \tab Pmisc\cr Type: \tab
  Package\cr Version: \tab 0.1\cr Date: \tab 2013-11-21\cr
  License: \tab GPL (>=2)\cr LazyLoad: \tab yes\cr }
}
\author{
  Paul Newcombe \email{paul.newcombe@mrc-bsu.cam.ac.uk}
}
\keyword{package}

