\name{Pmisc-package}
\alias{Pmisc-package}
\alias{Pmisc}
\docType{package}
\title{A collection of miscellaneous functions.}
\description{A collection of miscellaneous functions.}
\details{
\tabular{ll}{
  Package: \tab Pmisc\cr
  Type: \tab Package\cr
  Version: \tab 0.1-12-01-2015\cr
  License: \tab GPL (>=2)\cr }
A collection of miscellaneous functions.
}
\author{Paul Newcombe \email{paul.newcombe@mrc-bsu.cam.ac.uk}}
