\name{Pmisc-package}
\alias{Pmisc-package}
\alias{Pmisc}
\docType{package}
\title{All my miscellaneous functions.}
\description{All my miscellaneous functions.}
\details{
\tabular{ll}{
  Package: \tab Pmisc\cr
  Type: \tab Package\cr
  Version: \tab 0.1-19-07-2014\cr
  License: \tab GPL (>=2)\cr }
All my miscellaneous functions.
}
\author{Paul Newcombe \email{paul.newcombe@mrc-bsu.cam.ac.uk}}
