% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.R
\name{priorPostSd}
\alias{priorPostSd}
\alias{priorPost}
\alias{precToSd}
\title{precisions to standard deviations}
\usage{
priorPostSd(object, param = 1:length(object$all.hyper$random),
  group = c("random", "family"))
}
\arguments{
\item{object}{an inla result}

\item{param}{vector of parameters to transform}

\item{group}{random effect parameters or 'family' parameters}

\item{densmat}{matrix with columns 'x' and 'y'}
}
\description{
Transforms prior and posterior distributions of precision parameters to standard deviations
}
