% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgcpRaster.R
\name{spatialAtRisk.RasterLayer}
\alias{spatialAtRisk.RasterLayer}
\title{Convert rasters to lgcp objects}
\usage{
\method{spatialAtRisk}{RasterLayer}(X, ...)
}
\arguments{
\item{X}{an object of class \code{\link[raster]{RasterLayer-class}}}

\item{...}{not used}
}
\value{
A \code{fromXYZ} object produced by 
\code{\link[lgcp]{spatialAtRisk}}
}
\description{
Convert a RasterLayer to spatialAtRisk
}
\examples{
library('raster')
library('lgcp')
x = raster(matrix(1:12, 4,3,byrow=TRUE))/100
x2 = spatialAtRisk(x)
sum(values(x))*prod(res(x))
sum(x2$Zm)*prod(res(x))
par(mfrow=c(1,2))
plot(x)
plot(x2)
}

